% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{spark_read_tfrecord}
\alias{spark_read_tfrecord}
\title{Read a TFRecord File}
\usage{
spark_read_tfrecord(sc, name, path, schema = NULL,
  record_type = c("Example", "SequenceExample"), overwrite = TRUE)
}
\arguments{
\item{sc}{A spark conneciton.}

\item{name}{The name to assign to the newly generated table.}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports the "hdfs://", "s3a://" and "file://" protocols.}

\item{schema}{(Currently unsupported.) Schema of TensorFlow records.  If not provided, the schema is inferred from TensorFlow records.}

\item{record_type}{Input format of TensorFlow records. By default it is Example.}

\item{overwrite}{Boolean; overwrite the table with the given name if it already exists?}
}
\description{
Read a TFRecord file as a Spark DataFrame.
}
